+++ jest.spec.mjs
@@ -28,7 +28,7 @@ const config = {
     'node_modules/(?!@toptal|@topkit|d3|internmap|robust-predicates|delaunator)',
   ],
   coverageReporters: ['json', 'text-summary'],
-  // @TODO This option does not work due to bug in davinci-qa https://toptal-core.atlassian.net/browse/FX-4311
+  // TODO: This option does not work due to bug in davinci-qa https://toptal-core.atlassian.net/browse/FX-4311
   // collectCoverageFrom: ['packages/**/*.{ts,tsx}'],
 }