+++ packages/base/Autocomplete/src/Autocomplete/Autocomplete.tsx
@@ -95,7 +95,7 @@ export interface Props
   /** ReactNode for labels that will be used as end InputAdornment - */
   endAdornment?: ReactNode
   /**
-   * @deprecated Use the `status` prop instead to both support success and error states
+   * @deprecated [FX-4715] Use the `status` prop instead to both support success and error states
    * Indicate whether `Autocomplete` is in error state
    */
   error?: boolean
@@ -197,6 +197,7 @@ export const Autocomplete = forwardRef<HTMLInputElement, Props>(
       )
     }
 
+    // TODO: [FX-4715]
     usePropDeprecationWarning({
       props,
       name: 'error',