+++ packages/base/DatePicker/src/DatePicker/DatePicker.tsx
@@ -88,7 +88,7 @@ export interface Props
   /** Specify a value if want to enable browser autofill */
   autoComplete?: string
   /**
-   * @deprecated Use the `status` prop instead to both support success and error states
+   * @deprecated [FX-4715] Use the `status` prop instead to both support success and error states
    * Indicate whether `DatePicker` is in error state
    */
   error?: boolean
@@ -157,6 +157,7 @@ export const DatePicker = (props: Props) => {
   } = props
   const classes = useStyles()
 
+  // TODO: [FX-4715]
   usePropDeprecationWarning({
     props,
     name: 'error',