+++ packages/base/Dropzone/src/Dropzone/Dropzone.tsx
@@ -44,7 +44,7 @@ export interface Props extends BaseProps {
   /** Value uses the File interface. */
   value?: FileUpload[]
   /**
-   * @deprecated **Use value[n].error instead.**
+   * @deprecated [FX-4715] Use value[n].error instead.
    * Provide reasons why files couldn't be dropped into dropzone
    */
   errorMessages?: string[]
@@ -100,6 +100,7 @@ export const Dropzone = forwardRef<HTMLInputElement, Props>(function Dropzone(
 
   const classes = useStyles()
 
+  // TODO: [FX-4715]
   usePropDeprecationWarning({
     props,
     name: 'errorMessages',