+++ packages/base/FileInput/src/FileInput/index.ts
@@ -4,5 +4,5 @@ import type { Props as InternalFileInputProps } from './FileInput'
 export { default as FileInput } from './FileInput'
 export * from './types'
 export type FileInputProps = OmitInternalProps<InternalFileInputProps>
-/** @deprecated Use FileInputProps instead */
+/** @deprecated [FX-4714] Use FileInputProps instead */
 export type Props = FileInputProps