+++ packages/base/Input/src/Input/Input.tsx
@@ -46,7 +46,7 @@ export interface Props
   /** Placeholder for value */
   placeholder?: string
   /**
-   * @deprecated Use the `status` prop instead to both support success and error states
+   * @deprecated [FX-4715] Use the `status` prop instead to both support success and error states
    * Indicate whether `Input` is in error state
    */
   error?: boolean
@@ -283,6 +283,7 @@ export const Input = forwardRef<HTMLInputElement, Props>(function Input(
     ...rest
   } = purifyProps(props)
 
+  // TODO: [FX-4715]
   usePropDeprecationWarning({
     props,
     name: 'error',