+++ packages/base/Radio/src/RadioGroup/index.ts
@@ -3,3 +3,5 @@ import type { OmitInternalProps } from '@toptal/picasso-shared'
 import type { Props as InternalRadioGroupProps } from './RadioGroup'
 export { default as RadioGroup } from './RadioGroup'
 export type RadioGroupProps = OmitInternalProps<InternalRadioGroupProps>
+/** @deprecated [FX-4714] Use RadioGroupProps instead */
+export type Props = RadioGroupProps