+++ packages/base/Select/src/Select/Select.tsx
@@ -40,6 +40,7 @@ export const Select = documentable(
       { native, ...props }: SelectProps<T, M>,
       ref: React.Ref<HTMLInputElement> | null
     ) => {
+      // TODO: [FX-4715]
       usePropDeprecationWarning({
         props,
         name: 'error',
@@ -76,6 +77,7 @@ Select.defaultProps = {
   searchPlaceholder: 'Search',
   native: false,
   status: 'default',
+  enableResetSearch: false,
 }
 
 export default Select