+++ packages/base/Switch/src/Switch/index.ts
@@ -3,5 +3,5 @@ import type { OmitInternalProps } from '@toptal/picasso-shared'
 import type { Props as InternalSwitchProps } from './Switch'
 export { default as Switch } from './Switch'
 export type SwitchProps = OmitInternalProps<InternalSwitchProps>
-/** @deprecated Use SwitchProps instead */
+/** @deprecated [FX-4714] Use SwitchProps instead */
 export type Props = SwitchProps