+++ packages/base/Tagselector/src/TagSelector/TagSelector.tsx
@@ -59,7 +59,7 @@ export interface Props
   /** Disables `TagSelector` */
   disabled?: boolean
   /**
-   * @deprecated Use the `status` prop instead to both support success and error states
+   * @deprecated [FX-4715] Use the `status` prop instead to both support success and error states
    * Indicate whether `TagSelector` is in error state
    */
   error?: boolean
@@ -148,6 +148,7 @@ export const TagSelector = forwardRef<HTMLInputElement, Props>(
       ...rest
     } = props
 
+    // TODO: [FX-4715]
     usePropDeprecationWarning({
       props,
       name: 'error',