+++ packages/picasso-query-builder/src/QueryBuilder/story/Default.example.tsx
@@ -199,6 +199,8 @@ const fields: Field[] = [
     name: 'alsoPlays',
     label: 'Also plays',
     valueEditorType: 'multiselect',
+    enableResetSearch: true,
+    enableReset: true,
     values: musicalInstruments,
     defaultValue: 'More cowbell',
   },